Add HeroDetailComponent and update template
In this commit, a new HeroDetailComponent has been added. The new component includes a template file, a test file, and a corresponding TypeScript file. The AppComponent's module has been updated to import and declare the HeroDetailComponent. Also, the HeroesComponent's HTML file has been updated to use the new HeroDetailComponent. And finally, changed the name of one of the mock heroes in the mock-heroes.ts file.